refactor: name the CLI api module by domain and fold in the token call
